+++ empire/test/test_listener_api.py
@@ -7,7 +7,7 @@ def get_base_listener():
         "template": "http",
         "options": {
             "Name": "new-listener-1",
-            "Host": "http://localhost:1336",
+            "Host": "http://localhost",
             "BindIP": "0.0.0.0",
             "Port": "1336",
             "Launcher": "powershell -noP -sta -w 1 -enc ",
@@ -131,13 +131,13 @@ def test_create_listener_normalization_adds_protocol_and_default_port(
     base_listener = get_base_listener()
     base_listener["name"] = "temp123"
     base_listener["options"]["Host"] = "localhost"
-    base_listener["options"]["Port"] = None
+    base_listener["options"]["Port"] = "80"
 
     response = client.post(
         "/api/v2/listeners/", headers=admin_auth_header, json=base_listener
     )
     assert response.status_code == status.HTTP_201_CREATED
-    assert response.json()["options"]["Host"] == "http://localhost:80"
+    assert response.json()["host_address"] == "http://localhost/"
     assert response.json()["options"]["Port"] == "80"
 
     client.delete(
@@ -155,7 +155,7 @@ def test_create_listener_normalization_adds_port_to_host(client, admin_auth_head
         "/api/v2/listeners/", headers=admin_auth_header, json=base_listener
     )
     assert response.status_code == status.HTTP_201_CREATED
-    assert response.json()["options"]["Host"] == "http://localhost:1234"
+    assert response.json()["host_address"] == "http://localhost:1234/"
     assert response.json()["options"]["Port"] == "1234"
 
     client.delete(
@@ -174,13 +174,8 @@ def test_create_listener_normalization_preserves_user_defined_ports(
     response = client.post(
         "/api/v2/listeners/", headers=admin_auth_header, json=base_listener
     )
-    assert response.status_code == status.HTTP_201_CREATED
-    assert response.json()["options"]["Host"] == "http://localhost:443"
-    assert response.json()["options"]["Port"] == "1234"
-
-    client.delete(
-        f"/api/v2/listeners/{response.json()['id']}", headers=admin_auth_header
-    )
+    assert response.status_code == status.HTTP_400_BAD_REQUEST
+    assert response.json()["detail"] == "Port cannot be provided in a host name"
 
 
 def test_create_listener_normalization_sets_host_port_as_bind_port(
@@ -188,15 +183,14 @@ def test_create_listener_normalization_sets_host_port_as_bind_port(
 ):
     base_listener = get_base_listener()
     base_listener["name"] = "temp123"
-    base_listener["options"]["Host"] = "http://localhost:443"
-    base_listener["options"]["Port"] = None
+    base_listener["options"]["Host"] = "http://localhost"
+    base_listener["options"]["Port"] = "443"
 
     response = client.post(
         "/api/v2/listeners/", headers=admin_auth_header, json=base_listener
     )
     assert response.status_code == status.HTTP_201_CREATED
-    assert response.json()["options"]["Host"] == "http://localhost:443"
-    assert response.json()["options"]["Port"] == "443"
+    assert response.json()["host_address"] == "http://localhost:443/"
 
     client.delete(
         f"/api/v2/listeners/{response.json()['id']}", headers=admin_auth_header
@@ -208,16 +202,15 @@ def test_create_listener_with_https_host_no_cert_path(client, admin_auth_header)
     # because the listener might be behind a reverse proxy that handles TLS.
     base_listener = get_base_listener()
     base_listener["name"] = "temp123"
-    base_listener["options"]["Host"] = "https://localhost:443"
-    base_listener["options"]["Port"] = "80"
+    base_listener["options"]["Host"] = "https://localhost"
+    base_listener["options"]["Port"] = "443"
     base_listener["options"]["CertPath"] = ""
     response = client.post(
         "/api/v2/listeners/", headers=admin_auth_header, json=base_listener
     )
 
     assert response.status_code == status.HTTP_201_CREATED
-    assert response.json()["options"]["Host"] == "https://localhost:443"
-    assert response.json()["options"]["Port"] == "80"
+    assert response.json()["host_address"] == "https://localhost/"
     assert response.json()["options"]["CertPath"] == ""
 
     client.delete(